+++ _data/charts/benchmark_upload.json
@@ -51,97 +51,6 @@
       "values": []
     }
   ],
-  "data": [
-    {
-      "name": "code_uploads",
-      "values": [
-        {
-          "count": 17,
-          "name": "1b"
-        },
-        {
-          "count": 13,
-          "name": "1c"
-        },
-        {
-          "count": 12,
-          "name": "1a"
-        },
-        {
-          "count": 10,
-          "name": "3a"
-        },
-        {
-          "count": 8,
-          "name": "1d"
-        },
-        {
-          "count": 7,
-          "name": "2a"
-        },
-        {
-          "count": 6,
-          "name": "2b"
-        },
-        {
-          "count": 6,
-          "name": "2c"
-        },
-        {
-          "count": 6,
-          "name": "2d"
-        },
-        {
-          "count": 6,
-          "name": "7a"
-        },
-        {
-          "count": 5,
-          "name": "6a"
-        },
-        {
-          "count": 3,
-          "name": "6b"
-        },
-        {
-          "count": 1,
-          "name": "4a"
-        },
-        {
-          "count": 1,
-          "name": "4b"
-        },
-        {
-          "count": 1,
-          "name": "4c"
-        },
-        {
-          "count": 1,
-          "name": "4d"
-        },
-        {
-          "count": 1,
-          "name": "4e"
-        },
-        {
-          "count": 1,
-          "name": "4f"
-        },
-        {
-          "count": 1,
-          "name": "4g"
-        },
-        {
-          "count": 1,
-          "name": "4h"
-        },
-        {
-          "count": 1,
-          "name": "7b"
-        }
-      ]
-    }
-  ],
   "height": 115,
   "marks": [
     {
