--- _data/table.yaml.j2
@@ -1,14 +1,0 @@
----
-data:
-{%- for name, datum in data %}
-  - name: {{ name }}
-    code:
-      name: {{ datum.metadata.implementation.name }}
-      url: {{ datum.metadata.implementation.repo.url }}
-    id_: {{ datum.benchmark.id }}.{{ datum.benchmark.version }}
-    author:
-      email: {{ datum.metadata.author.email }}
-      name: {{ datum.metadata.author.first }} {{ datum.metadata.author.last }}
-    timestamp: {{ datum.metadata.timestamp | to_datetime(format_="%Y/%m/%d %H:%M:%S") }}
-    cores: {{ datum.metadata.hardware.cores }}
-{%- endfor %}
