+++ _includes/benchmarks.html
@@ -18,8 +18,7 @@
           </a>
           <h6>
             <br>
-            <a href="{{ site.baseurl }}/simulations/#simulations">
-              Total Uploads: {{ site.data.data_table.data | size }}
+            <a href="{{ site.baseurl }}/simulations/#simulations" id="total_uploads_bench">
             </a>
           </h6>
         </div>
