+++ js/uploads.coffee
@@ -9,3 +9,5 @@
 {% include coffee/uploads.coffee %}
 
 build_uploads(BENCHMARK_NUM, DATA, TAG)
+
+$('#total_uploads_bench').html('Total Uploads: ' + total_uploads(DATA))
