+++ simulations/index.html
@@ -28,8 +28,8 @@ comment: >-
 <div class="container">
   <div class="row small-buffer">
     <div class="col s12 m12 l12 xl8 offset-xl2">
-      <h5>
-        Total Uploads: {{ site.data.data_table.data | size }}
+      <h5 id="total_uploads">
+        <!-- Total Uploads: {{ site.data.simulations | size }} -->
       </h5>
     </div>
     <div class="col s12 m6 l5 xl4 offset-xl2">
@@ -44,7 +44,6 @@ comment: >-
 </div>
 
 <script src="{{ site.baseurl }}/js/upload_barchart.js"></script>
-<script src="{{ site.baseurl }}/js/simulation_chart.js"></script>
 
 {% include benchmarks.html upload=true %}
